Rename `boltdb` package to `boltprojection`.
