--- boltdb/doc.go
@@ -1,2 +1,0 @@
-// Package boltdb provides utilities for building BoltDB-based projections.
-package boltdb
