+++ boltprojection/bucket.go
@@ -1,4 +1,4 @@
-package boltdb
+package boltprojection
 
 import "go.etcd.io/bbolt"
