+++ boltprojection/doc.go
@@ -1,0 +1,3 @@
+// Package boltprojection provides utilities for building BoltDB-based
+// projections.
+package boltprojection
