+++ go.mod
@@ -3,6 +3,7 @@ module github.com/dogmatiq/projectionkit
 go 1.15
 
 require (
+	github.com/boltdb/bolt v1.3.1
 	github.com/dogmatiq/dogma v0.10.0
 	github.com/go-sql-driver/mysql v1.5.0
 	github.com/gofrs/uuid v3.3.0+incompatible // indirect
